fix(dedicated): account contact redirection (#5170)
ref: DTRSD-41441
